added esm support for server
