+++ server/constants.mjs
@@ -1,0 +1,4 @@
+export const SOCKET_REQUEST_USERNAME = "RequestUsername";
+export const SOCKET_USERNAME_OK = "UsernameOk";
+export const SOCKET_USERNAME_DENIED = "UsernameDenied";
+export const SOCKET_MESSAGE = "Message";
