--- src/constants.js
@@ -1,11 +1,0 @@
-const SOCKET_REQUEST_USERNAME = "RequestUsername";
-const SOCKET_USERNAME_OK = "UsernameOk";
-const SOCKET_USERNAME_DENIED = "UsernameDenied";
-const SOCKET_MESSAGE = "Message";
-
-module.exports = {
-  SOCKET_REQUEST_USERNAME,
-  SOCKET_USERNAME_OK,
-  SOCKET_USERNAME_DENIED,
-  SOCKET_MESSAGE,
-};
